Backend/requirements.txt
Flask
Flask-Cors
Pillow
ai-edge-litert
numpy
